+++ fabric/src/main/java/de/cech12/usefulhats/init/ModItems.java
@@ -42,7 +42,7 @@ public class ModItems {
     public static final Item STRAW_HAT = register("straw_hat", StrawHatItem::new);
     public static final Item WING_HELMET = register("wing_helmet", WingHelmetItem::new);
 
-    public static final DataComponentType<EnderHelmetItem.Position> ENDER_HELMET_POSITION = registerComponent("ender_helmet_position", (builder) -> builder.networkSynchronized(EnderHelmetItem.Position.STREAM_CODEC));
+    public static final DataComponentType<EnderHelmetItem.Position> ENDER_HELMET_POSITION = registerComponent("ender_helmet_position", (builder) -> builder.persistent(EnderHelmetItem.Position.CODEC).networkSynchronized(EnderHelmetItem.Position.STREAM_CODEC));
 
     static {
         Constants.ENDER_HELMET_POSITION = () -> ENDER_HELMET_POSITION;
@@ -57,7 +57,7 @@ public class ModItems {
     }
 
     private static <T> DataComponentType<T> registerComponent(String name, UnaryOperator<DataComponentType.Builder<T>> unaryOperator) {
-        return Registry.register(BuiltInRegistries.DATA_COMPONENT_TYPE, name, (unaryOperator.apply(DataComponentType.builder())).build());
+        return Registry.register(BuiltInRegistries.DATA_COMPONENT_TYPE, Constants.id(name), (unaryOperator.apply(DataComponentType.builder())).build());
     }
 
 }