+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.6-6.9.1.0] - 2025-10-13
+### Fixed
+- location of Ender Helmet was not saved persistently (thanks to Gunman1313 for the report) #76
+
 ## [1.21.6-6.9.0.0] - 2025-06-19
 ### Changed
 - Updated to Minecraft 1.21.6 (Fabric 0.127.1+1.21.6, NeoForge 21.6.6-beta, Forge 56.0.3)